Automatically use ES6 if Babel Transpiler is found
Use the Babel::Transpiler to compile all javascript if found and fall
back to plain javascript.

This also normalizes the lang attribute, so that a missing attribute,
the value 'javascript' and the value 'es6' (if Babel is found) are
handled the same and it fails if the lang(uage) is not supported.